backend/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
faiss-cpu
requests
PyPDF2
numpy
pydantic
python-multipart
